Converting Mapper class into processors - HZN-766
